Consolidate naming of GradleModuleServices implementations
