+++ platforms/core-configuration/declarative-dsl-provider/src/main/resources/META-INF/services/org.gradle.internal.service.scopes.GradleModuleServices
@@ -1,1 +1,1 @@
-org.gradle.internal.declarativedsl.provider.DeclarativeDslServiceRegistry
+org.gradle.internal.declarativedsl.provider.DeclarativeDslServices
