+++ platforms/core-configuration/kotlin-dsl-provider-plugins/src/main/resources/META-INF/services/org.gradle.internal.service.scopes.GradleModuleServices
@@ -1,1 +1,1 @@
-org.gradle.kotlin.dsl.provider.plugins.KotlinDslProviderPluginsServiceRegistry
+org.gradle.kotlin.dsl.provider.plugins.KotlinDslProviderServices
