+++ platforms/core-configuration/kotlin-dsl/src/main/resources/META-INF/services/org.gradle.internal.service.scopes.GradleModuleServices
@@ -1,1 +1,1 @@
-org.gradle.kotlin.dsl.services.KotlinScriptServiceRegistry
+org.gradle.kotlin.dsl.services.KotlinScriptServices
