+++ platforms/extensibility/plugin-use/src/main/resources/META-INF/services/org.gradle.internal.service.scopes.GradleModuleServices
@@ -1,1 +1,1 @@
-org.gradle.plugin.internal.PluginUsePluginServiceRegistry
+org.gradle.plugin.internal.PluginUseServices
