+++ platforms/jvm/language-java/src/main/java/org/gradle/language/java/internal/JavaToolchainServices.java
@@ -30,7 +30,7 @@ import org.gradle.process.internal.worker.child.WorkerDirectoryProvider;
 import org.gradle.workers.internal.ActionExecutionSpecFactory;
 import org.gradle.workers.internal.WorkerDaemonFactory;
 
-public class JavaToolChainServiceRegistry extends AbstractGradleModuleServices {
+public class JavaToolchainServices extends AbstractGradleModuleServices {
     @Override
     public void registerProjectServices(ServiceRegistration registration) {
         registration.addProvider(new ProjectScopeCompileServices());
