+++ platforms/jvm/language-java/src/main/resources/META-INF/services/org.gradle.internal.service.scopes.GradleModuleServices
@@ -1,3 +1,3 @@
 org.gradle.api.internal.tasks.CompileServices
-org.gradle.language.java.internal.JavaToolChainServiceRegistry
-org.gradle.language.java.internal.JavaLanguagePluginServiceRegistry
+org.gradle.language.java.internal.JavaToolchainServices
+org.gradle.language.java.internal.JavaLanguageServices
