+++ platforms/jvm/language-jvm/src/main/resources/META-INF/services/org.gradle.internal.service.scopes.GradleModuleServices
@@ -1,1 +1,1 @@
-org.gradle.language.jvm.internal.JvmPluginServiceRegistry
+org.gradle.language.jvm.internal.JvmLanguageServices
