+++ platforms/jvm/plugins-java-base/src/main/resources/META-INF/services/org.gradle.internal.service.scopes.GradleModuleServices
@@ -1,1 +1,1 @@
-org.gradle.api.plugins.internal.JvmLanguageUtilsServiceRegistry
+org.gradle.api.plugins.internal.JvmLanguageUtilsServices
