+++ platforms/software/platform-base/src/main/resources/META-INF/services/org.gradle.internal.service.scopes.GradleModuleServices
@@ -1,1 +1,1 @@
-org.gradle.platform.base.internal.registry.ComponentModelBaseServiceRegistry
+org.gradle.platform.base.internal.registry.ComponentModelBaseServices
