+++ platforms/software/resources-gcs/src/main/resources/META-INF/services/org.gradle.internal.service.scopes.GradleModuleServices
@@ -1,1 +1,1 @@
-org.gradle.internal.resource.transport.gcp.gcs.GcsResourcesPluginServiceRegistry
+org.gradle.internal.resource.transport.gcp.gcs.GcsResourcesServices
