+++ platforms/software/resources-http/src/main/java/org/gradle/internal/resource/transport/http/HttpResourcesServices.java
@@ -28,7 +28,7 @@ import org.gradle.internal.resource.connector.ResourceConnectorFactory;
 import org.gradle.internal.service.ServiceRegistration;
 import org.gradle.internal.service.scopes.AbstractGradleModuleServices;
 
-public class HttpResourcesPluginServiceRegistry extends AbstractGradleModuleServices {
+public class HttpResourcesServices extends AbstractGradleModuleServices {
     @Override
     public void registerGlobalServices(ServiceRegistration registration) {
         registration.addProvider(new GlobalScopeServices());
