+++ platforms/software/resources-http/src/main/resources/META-INF/services/org.gradle.internal.service.scopes.GradleModuleServices
@@ -1,2 +1,2 @@
-org.gradle.internal.resource.transport.http.HttpResourcesPluginServiceRegistry
+org.gradle.internal.resource.transport.http.HttpResourcesServices
