+++ platforms/software/resources-s3/src/main/resources/META-INF/services/org.gradle.internal.service.scopes.GradleModuleServices
@@ -1,2 +1,2 @@
-org.gradle.internal.resource.transport.aws.s3.S3ResourcesPluginServiceRegistry
+org.gradle.internal.resource.transport.aws.s3.S3ResourcesServices
