+++ platforms/software/resources-sftp/src/main/resources/META-INF/services/org.gradle.internal.service.scopes.GradleModuleServices
@@ -1,2 +1,2 @@
-org.gradle.internal.resource.transport.sftp.SftpResourcesPluginServiceRegistry
+org.gradle.internal.resource.transport.sftp.SftpResourcesServices
