+++ platforms/software/testing-base/src/main/resources/META-INF/services/org.gradle.internal.service.scopes.GradleModuleServices
@@ -1,1 +1,1 @@
-org.gradle.api.internal.tasks.testing.TestingBasePluginServiceRegistry
+org.gradle.api.internal.tasks.testing.TestingBasePluginServices
